feat:Initial scripts for LED and switch.
These are the initial scripts to turn the LED on and off.
Also the ability to detect the switch press.
Adding the sounds that were used to test and the sound for the
Emergency call activation.
The config files are the differences that needed to be made to the
kernel to be able to get SND_USB into the kernel.

 On branch main
 Initial commit
 Changes to be committed:
	new file:   TF-config
	new file:   digi-config
	new file:   led_blue.sh
	new file:   led_green.sh
	new file:   led_red.sh
	new file:   sounds/ENU00012.wav
	new file:   sounds/ENU00209.wav
	new file:   sounds/ENU00439.wav
	new file:   sounds/S0000209.wav
	new file:   sounds/S0000300.wav
	new file:   sounds/S0000301.wav
	new file:   sounds/S0000302.wav
	new file:   sounds/S0000303.wav
	new file:   sounds/S0000304.wav
	new file:   sounds/S0000305.wav
	new file:   sounds/record.wav
	new file:   start.sh
	new file:   switch_detect.sh
	new file:   switch_mon.sh