fix: Removed extension from place_call.py
The code to send the info packet to the EDC server was refactored out
of this code to create a stand alone script to be able to send other
types of info packets. The extension argument was removed from
place_call.py be cause it was not needed.
As a result the extension was removed from this script.
 On branch br_EDC_code_refactor
 Changes to be committed:
	modified:   VOIP/voip_call_monitor_tcp.py