feat:Handle collision of in and out calls.
In the case of the inbound call not being in the whitelist,
the incoming call is sent a AT+CHUP.
However, there could still be a collision between the dial out
and the phone begining to ring.
The status is now reported as
```
Response Status: pending
Message: Modem busy in CALL_ENDING state, command queued
```
Which will allow the calling code to retry if necessary.

 On branch br_place_call_morph
 Changes to be committed:
	modified:   manage_modem.py