feat:Program to handle incoming and outgoing calls
This is a rework of place_call.py (which is still needed because it is
being sourced) to handle both the placing of outbound calls, and
handling inbound calls.

The inbound calls complete and the caller is in the conference as
seen by the asterisk cli. However, the audio and dtmf don't appear to be
making it all the way.
 On branch br_modem_manage
 Changes to be committed:
	modified:   manage_modem.py